[refactor] fetch clarify and translate from recap module
