+++ app/telegram/__init__.py
@@ -1,1 +1,10 @@
-from . import recap, note, study, note_list, language, onboarding
+from . import (
+    recap,
+    note,
+    study,
+    note_list,
+    language,
+    onboarding,
+    translate,
+    clarify,
+)
